perf(backend): change backend struct
