--- backend/client/main.go
@@ -1,11 +1,0 @@
-/*
-Copyright © 2022 NAME HERE <EMAIL ADDRESS>
-
-*/
-package main
-
-import "github.com/KenethSandoval/doc-md/client/cmd"
-
-func main() {
-	cmd.Execute()
-}
