+++ backend/cmd/cli/commands/root.go
@@ -1,0 +1,24 @@
+package commands
+
+import (
+	"os"
+
+	"github.com/spf13/cobra"
+)
+
+var rootCmd = &cobra.Command{
+	Use:   "cligrpc",
+	Short: "A client for grpc server",
+	Long:  `A client for grpc server, send a yml data file`,
+}
+
+func Execute() {
+	err := rootCmd.Execute()
+	if err != nil {
+		os.Exit(1)
+	}
+}
+
+func init() {
+	rootCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+}
