+++ backend/cmd/cli/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/KenethSandoval/doc-md/cmd/cli/commands"
+
+func main() {
+	commands.Execute()
+}
