+++ backend/docker-compose.yml
@@ -1,0 +1,9 @@
+version: "3.3"
+
+services:
+  mongo:
+    image: mongo
+    ports:
+      - 27017:27017
+    volumes:
+      - ./mongo_data:/data/db
